+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 + Added area mini bosses spawn timer.
 + Added Corleone highlighter.
 + Added **Tps Display** - Show the Tps of the current server.
++ Added Kuudra Key Number overlay.
++ Added colored highlight support for zealots, bruisers and special zealots.
 
 ## Removals
 - Removed Blaze slayer Pillar warning + timer (The Feature caused lags and Soopy's approach is better)